Rearrange project structure to separate database from use case logic
