+++ src/ServiceAssessmentService/ServiceAssessmentService.WebApp/Pages/Book/Create.cshtml.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using ServiceAssessmentService.Application;
+using ServiceAssessmentService.Application.UseCases;
 
 namespace ServiceAssessmentService.WebApp.Pages.Book;
