+++ src/ServiceAssessmentService/ServiceAssessmentService.WebApp/Pages/Book/View.cshtml.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using ServiceAssessmentService.Application;
+using ServiceAssessmentService.Application.UseCases;
 using ServiceAssessmentService.Domain.Model;
 
 namespace ServiceAssessmentService.WebApp.Pages.Book;
