+++ src/ServiceAssessmentService/ServiceAssessmentService.WebApp/Pages/Shared/_DfE_1200px_Layout.cshtml
@@ -1,5 +1,4 @@
 ﻿@using Microsoft.AspNetCore.Identity
-@using ServiceAssessmentService.Application.Entities
 @using GovUk.Frontend.AspNetCore.TagHelpers
